chore(tests): move tests to the root
