+++ tests/Console/UrlGeneratorTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Imdhemy\Purchases\Tests\Unit\Console;
+namespace Imdhemy\Purchases\Tests\Console;
 
 use Illuminate\Foundation\Bootstrap\LoadConfiguration as IlluminateLoadConfiguration;
 use Illuminate\Http\Request;
