+++ tests/ServerNotifications/AppStoreServerNotificationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Imdhemy\Purchases\Tests\Unit\ServerNotifications;
+namespace Imdhemy\Purchases\Tests\ServerNotifications;
 
 use Imdhemy\AppStore\ServerNotifications\ServerNotification;
 use Imdhemy\Purchases\Contracts\SubscriptionContract;
