+++ tests/ServerNotifications/AppStoreV2ServerNotificationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Imdhemy\Purchases\Tests\Unit\ServerNotifications;
+namespace Imdhemy\Purchases\Tests\ServerNotifications;
 
 use Imdhemy\AppStore\ServerNotifications\V2DecodedPayload;
 use Imdhemy\Purchases\ServerNotifications\AppStoreV2ServerNotification;
