+++ tests/ServerNotifications/GoogleServerNotificationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Imdhemy\Purchases\Tests\Unit\ServerNotifications;
+namespace Imdhemy\Purchases\Tests\ServerNotifications;
 
 use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\Psr7\Response;
