+++ tests/ValueObjects/TimeTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Imdhemy\Purchases\Tests\Unit\ValueObjects;
+namespace Imdhemy\Purchases\Tests\ValueObjects;
 
 use Carbon\Carbon;
 use DateTime;
